+++ src/entidades/Proveedor.java
@@ -2,7 +2,7 @@ package entidades;
 
 public class Proveedor {
     private int idProveedor;
-    private int cuit;
+    private Long cuit;
     private String razonSocial;
     private String domicilio;
     private String telefono;
@@ -11,7 +11,7 @@ public class Proveedor {
     public Proveedor() {
     }
 
-    public Proveedor(int idProveedor, int cuit, String razonSocial, String domicilio, String telefono) {
+    public Proveedor(int idProveedor, Long cuit, String razonSocial, String domicilio, String telefono) {
         this.idProveedor = idProveedor;
         this.cuit = cuit;
         this.razonSocial = razonSocial;
@@ -19,7 +19,7 @@ public class Proveedor {
         this.telefono = telefono;
     }
     
-    public Proveedor(int cuit, String razonSocial, String domicilio, String telefono) {
+    public Proveedor(Long cuit, String razonSocial, String domicilio, String telefono) {
         this.cuit = cuit;
     	this.razonSocial = razonSocial;
         this.domicilio = domicilio;
@@ -30,9 +30,9 @@ public class Proveedor {
 
 	public void setIdProveedor(int idProveedor) {this.idProveedor = idProveedor;}
 	
-	public int getCuit() {return this.cuit;}
+	public Long getCuit() {return this.cuit;}
 	
-	public void setCuit(int cuit) {this.cuit = cuit;}
+	public void setCuit(Long cuit) {this.cuit = cuit;}
 
 	public String getRazonSocial() {return this.razonSocial;}
